Accept optional GlobalSettings
