+++ Sources/CharcoalSwiftUI/Extensions/EnvironmentObject+Extension.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+extension EnvironmentObject {
+    var hasValue: Bool {
+        !String(describing: self).contains("_store: nil")
+    }
+}
